% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{active}
\alias{active}
\title{List active instances}
\usage{
active()
}
\value{
a list of active instances
}
\description{
A list containing the current sessions active instances.
}
\examples{
bb <- listen(beakr = beakr(), daemonized = TRUE)
active()
}
