% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{addListener}
\alias{addListener}
\title{Internal function to add listeners}
\usage{
addListener(beakr, event, FUN)
}
\arguments{
\item{beakr}{a beakr instance}

\item{event}{an event (\emph{"start", "finish", "error"})}

\item{FUN}{middleware function.}
}
\description{
Internal function to add listeners
}
\keyword{internal}
