% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addMiddleware}
\alias{addMiddleware}
\title{Internal function to add middleware}
\usage{
addMiddleware(beakr, FUN, path = NULL, method = NULL,
  websocket = FALSE)
}
\arguments{
\item{websocket}{}
}
\description{
Internal function to add middleware
}
