% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beakr.R
\docType{package}
\name{Beakr}
\alias{Beakr}
\alias{Beakr-package}
\title{A Simple Web Framework for R.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Beakr
}
\description{
A web framework mimicking the functionality found in python Flask
including basic functionality for handling web GET and PUT requests. This
package is a ground-up rewrite of the original "jug" package by Bart Smeets.
}
\keyword{datasets}
