% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createBeakr}
\alias{createBeakr}
\title{Create a new Beakr instance}
\usage{
createBeakr(name = NULL)
}
\arguments{
\item{name}{an optional name assigned to the \emph{Beakr} object.}
}
\description{
Create a \code{Beakr} instance object by calling the top-level
\code{createBeakr()} function. If \code{name} is not supplied, a random name
will be assigned.
}
