% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{delete}
\alias{delete}
\title{DELETE-binding middleware}
\usage{
delete(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP DELETE requests to the specified path with the specified callback
functions or middleware.
}
\examples{
\dontrun{
# Create a beakr instance
server <- beakr() \%>\%
  delete("/", function(req, res, err) { "Successful DELETE request!" })
# Listen for HTTP/WebSocket requests
listen(server, daemonized = TRUE)
# In terminal:
$ curl -X DELETE http://127.0.0.1:8080/
Successful DELETE request!
}
}
