% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.jsonError}
\alias{.jsonError}
\title{JSON error function}
\usage{
.jsonError(req, res, err)
}
\value{
The incoming \code{res} object is modified.
}
\description{
This function is used to add a JSON error response to the
\code{res} object. It is called by the \code{handleErrors()} utility
function.
}
\keyword{internal}
