% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.matchPath}
\alias{.matchPath}
\title{Regex path query}
\usage{
.matchPath(pattern, path, ...)
}
\arguments{
\item{pattern}{String pattern to parse. (A middleware path.)}

\item{path}{Path to match to. (A request path.)}

\item{...}{Additional parameters passed to \code{regexpr()}.}
}
\value{
A List with information on matching paths and URL parameters.
}
\description{
This function is used in the \code{Router$invoke()}
method to match middleware paths to request paths.
}
\keyword{internal}
