% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.parseParameters}
\alias{.parseParameters}
\title{Parse the parameters passed by in the request}
\usage{
.parseParameters(req, body, query, type)
}
\arguments{
\item{req}{HTTP request object.}

\item{body}{Body text string.}

\item{query}{Url-encoded query string.}

\item{type}{Media mime type.}
}
\value{
A list of parameters and values.
}
\description{
Internal function used in the \code{Request$initialize()}
method to extract URL parameters from the request path.
}
\keyword{internal}
