% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.routeMiddleware}
\alias{.routeMiddleware}
\title{Internal function to add middleware to a Beakr object}
\usage{
.routeMiddleware(beakr, FUN, path = NULL, method = NULL, websocket = FALSE)
}
\arguments{
\item{beakr}{Beakr instance.}

\item{FUN}{Function to route middleware.}

\item{path}{Path to route the middleware.}

\item{method}{HTTP method to employ.}

\item{websocket}{Boolean, TRUE if websocket.}
}
\value{
A \code{Beakr} object with added middleware.
}
\description{
This function is used in each of the \code{http_~()} methods
as well as the \code{cors()}, \code{websocket()} and \code{use()} utility
functions.
}
\keyword{internal}
