% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{get}
\alias{get}
\title{GET-binding middleware}
\usage{
get(beakr, path = NULL, ...)
}
\arguments{
\item{beakr}{a beakr instance or \code{NULL}.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP GET requests to the specified path with the specified callback
functions or middleware.
}
\examples{
\dontrun{
# Create a beakr instance
server <- newBeakr() \%>\%
  get(path = "/", function(req, res, err) { "Hello, world!" })
# Listen for HTTP/WebSocket requests
listen(server)
# In terminal:
#  curl http://127.0.0.1:8080/
# > Hello, world!
}
}
