% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handleErrors}
\alias{handleErrors}
\title{Error handling middleware for a beakr instance}
\usage{
handleErrors(beakr, path = NULL)
}
\arguments{
\item{beakr}{Beakr instance}

\item{path}{Path for which the middleware is invoked, typically \code{NULL}.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
This default error-handling middleware function should be added
at the end of the middleware function pipeline. Any errors will be returned
within a JSON wrapper.
}
