% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{http_delete}
\alias{http_delete}
\title{DELETE-binding middleware}
\usage{
http_delete(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
Routes HTTP DELETE requests to the specified path with the
specified callback functions or middleware.
}
