% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{http_get}
\alias{http_get}
\title{GET-binding middleware}
\usage{
http_get(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance or \code{NULL}.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
Routes HTTP GET requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
new_beakr() \%>\%
  http_get("/", function(req, res, err) {
    return("Successful GET request!\n")
  }) \%>\%
  listen()

# In browser:
#   http://127.0.0.1:8080
# > Successful GET request!
}

}
