% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{http_post}
\alias{http_post}
\title{POST-binding middleware}
\usage{
http_post(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
Routes HTTP POST requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
newBeakr() \%>\%
  http_post("/", function(req, res, err) {
    return("Successful POST request!\n")
  }) \%>\%
  listen()

# In terminal:
#  curl -X POST http://127.0.0.1:8080/
# > Successful POST request!
}
}
