% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{http_put}
\alias{http_put}
\title{PUT-binding middleware}
\usage{
http_put(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
Routes HTTP PUT requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
new_beakr() \%>\%
  http_put("/", function(req, res, err) {
    return("Successful PUT request!\n")
  }) \%>\%
  listen()

# In terminal:
#  curl -X PUT http://127.0.0.1:8080/
# > Successful PUT request!
}

}
