% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{httpPUT}
\alias{httpPUT}
\title{PUT-binding middleware}
\usage{
httpPUT(beakr, path = NULL, FUN = NULL)
}
\arguments{
\item{beakr}{\code{Beakr} instance or \code{NULL}.}

\item{path}{String representing a path for which the middleware
function is invoked.}

\item{FUN}{Middleware function to be invoked.}
}
\value{
A \code{Beakr} object with added middleware.
}
\description{
Routes HTTP PUT requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
# Create an new Beakr instance
beakr <- newBeakr()

# Create a simple beakr pipeline
beakr \%>\%
  httpPUT("/", function(req, res, err) {
    return("Successful PUT request!\n")
  }) \%>\%
  listen(host = '127.0.0.1', port = 25118, daemon = TRUE)

# ------------------------------------------------------------
# IN A TERMINAL:
#  curl -X PUT http://127.0.0.1:25118/
# > Successful PUT request!
# ------------------------------------------------------------

# Kill the beakr instance
kill(beakr)
}
}
