% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{httpuv_stopAllServers}
\alias{httpuv_stopAllServers}
\title{Stop all servers}
\usage{
httpuv_stopAllServers()
}
\value{
None
}
\description{
Stops all \code{Beakr} instances (and any other servers created
with the \code{httpuv} package). This convenience function is a wrapper for
\code{httpuv::stopAllServers()} and is included to encourage experimentation
so that users who create multiple \code{Beakr} instances can quickly stop
them all.
}
\examples{
beakr1 <- newBeakr()
beakr2 <- newBeakr()
beakr1 \%>\% listen(daemon = TRUE, port = 1234, verbose = TRUE)
beakr2 \%>\% listen(daemon = TRUE, port = 4321, verbose = TRUE)
length(httpuv_listServers())
httpuv_stopAllServers()
length(httpuv_listServers())
}
