% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{include}
\alias{include}
\title{Construct an instance}
\usage{
include(beakr, include, file = NULL)
}
\arguments{
\item{beakr}{primary beakr instance.}

\item{include}{the external middleware to include.}

\item{file}{the source file path, if external middleware in separate .R file.}
}
\description{
The \code{include} function is used to merge beakr middleware
that has been constructed elsewhere.
}
\details{
Hierarchy
}
