% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-middleware.R
\name{jsonError}
\alias{jsonError}
\title{JSON error function}
\usage{
jsonError(req, res, err)
}
\arguments{
\item{req}{\code{Request} object.}

\item{res}{\code{Response} object.}

\item{err}{\code{Error} Error object.}
}
\value{
The incoming \code{res} object is modified.
}
\description{
This function is used to add a JSON error response to the
\code{res} object. It is called by the \code{handleErrors()} utility
function.
}
\seealso{
\link{Request}, \link{Response}, \link{Error}
}
