% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kill}
\alias{kill}
\title{Stop a beakr instance}
\usage{
kill(beakr, verbose = TRUE)
}
\arguments{
\item{beakr}{\code{Beakr} instance.}

\item{verbose}{Logical specifying whether to print out details of the
\code{Beakr} instance just stopped.}
}
\value{
None
}
\description{
Stops an active \code{Beakr} instance, closing
all open connections and unbinding the port.
}
\examples{
beakr <- newBeakr()
beakr \%>\%
  listen(daemon = TRUE, verbose = TRUE)
kill(beakr, verbose = TRUE)
}
