% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kill_all}
\alias{kill_all}
\title{Stop all beakr instances}
\usage{
kill_all()
}
\value{
None
}
\description{
Stops all instances that have been activated by
\code{\link{listen}} in the session.
}
\examples{
\dontrun{
new_beakr() \%>\% listen(daemon = TRUE)
# Kill all instances
kill_all()
## Stopped All Instances
}
}
\seealso{
\code{\link{kill}} and \code{\link{listen}}
}
