% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{killAll}
\alias{killAll}
\title{Stop all beakr instances}
\usage{
killAll()
}
\description{
Stops all instances that have been activated by
\code{\link{listen}} in the session.
}
\seealso{
\code{\link{kill}} and \code{\link{listen}}
}
