% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_active}
\alias{list_active}
\title{List active instances}
\usage{
list_active()
}
\value{
a list of active instances

A `list` of all active instances
}
\description{
A list containing the current sessions active instances.
}
\examples{
\donttest{
beakr('Test1') \%>\% listen(port = 1234, daemon = TRUE)
beakr('Test2') \%>\% listen(port = 2234, daemon = TRUE)
list_active()
## [[1]]
## [1] "Host: 127.0.0.1 | Port: 1234 | Active: TRUE"

## [[2]]
## [1] "Host: 127.0.0.1 | Port: 2234 | Active: TRUE"
}
}
