% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{Instance logging}
\usage{
logger(beakr, level = "DEBUG", file = NULL, echo = TRUE)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{level}{the log level (i.e. 'DEBUG', 'INFO', etc.).}

\item{file}{if provided, a file to write log output to.}

\item{echo}{if TRUE will print the log to console.}
}
\description{
A wrapper for \emph{futile.logger} to log events.
Currently supported logged events: \emph{'start', 'finish', error'}.
}
