% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{matchPath}
\alias{matchPath}
\title{Internal regex path query.}
\usage{
matchPath(pattern, path, ...)
}
\arguments{
\item{pattern}{the string pattern to parse}

\item{path}{the path to match to}

\item{...}{additional parameters}
}
\description{
Internal regex path query.
}
