% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaJug.R
\docType{package}
\name{MazamaJug}
\alias{MazamaJug}
\alias{MazamaJug-package}
\title{A Simple Web Framework for R.}
\description{
A web framework mimicking the functionality found in python Flask
including basic functionality for handling web GET and PUT requests. This
package is a ground-up rewrite of the original "jug" package by Bart Smeets.
}
