% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Middleware.R
\docType{data}
\name{Middleware}
\alias{Middleware}
\title{Middleware Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Middleware
}
\description{
Middleware Class
}
\keyword{datasets}
