% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_beakr}
\alias{new_beakr}
\title{Create a new Beakr instance}
\usage{
new_beakr(name = NULL)
}
\arguments{
\item{name}{an optional name assigned to the \emph{Beakr} object.}
}
\value{
A new and empty `beakr` App.
}
\description{
Create a \emph{Beakr} instance object by calling the top-level
\code{new_beakr()} function. If \code{name} is not supplied, a random name
will be assigned.
}
\examples{
\dontrun{
# Standard
beakr <- new_beakr()
listen(beakr)
# Pipe
new_beakr() \%>\% listen()
}
}
