% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onEvent}
\alias{onEvent}
\title{Beakr Event Listener}
\usage{
onEvent(beakr, event, FUN)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{event}{the event to listen for, (\emph{"start", "finish", "end"}).}

\item{FUN}{the response middleware function.}
}
\description{
Add an event listener to a \emph{beakr} instance. Currently
supported events are \code{"start", "finish", "error"}. The events \code{"start"} and
\code{"finish"} will pass the current state of the \code{req}, \code{res} and \code{err}
objects to the Listener. The \code{"error"} event will pass string error message.
}
