% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseParameters}
\alias{parseParameters}
\title{Parse the parameters passed by the request}
\usage{
parseParameters(request, body, query, type)
}
\arguments{
\item{type}{}
}
\value{
parsed parameters list
}
\description{
Parse the parameters passed by the request
}
