% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{post}
\alias{post}
\title{POST-binding middleware}
\usage{
post(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP POST requests to the specified path with the specified callback
functions or middleware.
}
\examples{
\dontrun{
# Create a beakr instance
server <- newBeakr() \%>\%
  post("/", function(req, res, err) { "Successful POST request!" })
# Listen for HTTP/WebSocket requests
listen(server)
# In terminal:
#  curl -X POST http://127.0.0.1:8080/
# > Successful POST request!
}
}
