% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{postr}
\alias{postr}
\title{POST-binding middleware}
\usage{
postr(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP postr requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
createBeakr() \%>\%
  postr("/", function(req, res, err) {
    return("Successful postr request!\n")
  }) \%>\%
  listen()

# In terminal:
#  curl -X postr http://127.0.0.1:8080/
# > Successful postr request!
}
}
