% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestRequest.R
\name{processTestRequest}
\alias{processTestRequest}
\title{Initialize process of test request}
\usage{
processTestRequest(beakr, test_request)
}
\arguments{
\item{beakr}{the beakr instance}

\item{test_request}{the TestRequest instance}
}
\description{
Initialize process of test request
}
