% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{PUT}
\alias{PUT}
\title{PUT-binding middleware}
\usage{
PUT(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP PUT requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
newBeakr() \%>\%
  PUT("/", function(req, res, err) {
    return("Successful PUT request!\\n")
  }) \%>\%
  startBeakr()

# In terminal:
#  curl -X PUT http://127.0.0.1:8080/
# > Successful PUT request!
}

}
