% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{putr}
\alias{putr}
\title{PUT-binding middleware}
\usage{
putr(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP putr requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
createBeakr() \%>\%
  putr("/", function(req, res, err) {
    return("Successful putr request!\n")
  }) \%>\%
  listen()

# In terminal:
#  curl -X putr http://127.0.0.1:8080/
# > Successful putr request!
}

}
