% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestHandler.R
\docType{data}
\name{RequestHandler}
\alias{RequestHandler}
\title{Request Handler Object Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RequestHandler
}
\description{
Request Handler Object Class
}
\keyword{datasets}
