% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{routeMiddleware}
\alias{routeMiddleware}
\title{Internal function to add middleware}
\usage{
routeMiddleware(beakr, FUN, path = NULL, method = NULL,
  websocket = FALSE)
}
\arguments{
\item{websocket}{}
}
\description{
Internal function to add middleware
}
