% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serveStaticFiles}
\alias{serveStaticFiles}
\title{Serve static files}
\usage{
serveStaticFiles(beakr = NULL, route = NULL, filePath = NULL)
}
\arguments{
\item{beakr}{\code{Beakr} instance or \code{NULL}.}

\item{route}{String representing a \emph{route}, a virtual directory path
where the static file will appear in a URL.}

\item{filePath}{String representing path and name of the file to be served
relative to the directory in which the beakr script is running.}
}
\value{
A \code{Beakr} with added middleware.
}
\description{
Binds to GET requests that aren't handled by specified paths.
Should support all filetypes; returns image and octet-stream types as a raw
string.
}
\details{
Serve static files from the host machine. Currently supports images,
PDF, JSON, HTML, and raw text. The static file appear at a URL composed of
the \code{route} and the file name and extension which will be extracted
from \code{filePath}.

For example, specifying:
\preformatted{
beakr %>%
  ...
  serverStaticFiles(/)
  ...
  listen()
}

TODO
}
