% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startBeakr}
\alias{startBeakr}
\title{Start a beakr instance and listen for connections}
\usage{
startBeakr(beakr, host, port, daemon, verbose)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{host}{a string that is a valid IPv4 or IPv6 address to listen on.}

\item{port}{a number or integer that indicates the port to listen on.}

\item{daemon}{run the instance in the background.}

\item{verbose}{boolean, debugging.}
}
\description{
Binds and listens for connections on the specified host and port.
}
\details{
\code{startBeakr()} binds the port and listens for connections on a thread.
The thread handles the I/O, and when it receives a HTTP request, it
will schedule a call to the user-defined middleware and handle the
request.

If the daemon boolean value is set to \code{TRUE}, \code{startBeakr()} binds
the specified port and listens for connections on an thread running in the
background.

See the \code{httpuv} package for more information.
}
\examples{
\dontrun{
newBeakr() \%>\%
  GET("/", function(req, res, err) {
    return("Successful GET request!\\n")
  }) \%>\%
  startBeakr()
}

}
