% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{static}
\alias{static}
\title{Serve static files}
\usage{
static(beakr, path = NULL, file = NULL)
}
\arguments{
\item{beakr}{a beakr instance}

\item{path}{a string representing a relative path for which the middleware
is invoked.}

\item{file}{a string representing the path to the file that is to be served.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
Binds to GET requests that aren't handled by specified paths.
Should support all filetypes; returns image and octet-stream types as a raw
string.
}
\details{
Serve static files from the host machine. Currently supports images,
PDF, JSON, HTML, and raw text. The static file will be served on the `path`
specified in addition to the file name and extension. For example, an image
`example_dir/path_to_pictures/pic.png` will be served on the URL extension
`/path/pic.png`.
}
\examples{
\dontrun{
path_to_file <- "example_dir/path_to_pictures/pic.png"
newBeakr() \%>\%
  static(path = 'readme/', file = path_to_file) \%>\%
  listen()
}
}
