% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopAllServers}
\alias{stopAllServers}
\title{Stop all servers}
\usage{
stopAllServers()
}
\value{
None
}
\description{
Stops all \code{Beakr} servers currently running (and any other
servers created with the \pkg{httpuv} package). This function is included to
encourage experimentation so that users who create multiple \code{Beakr}
instances can quickly find and stop them all.

See \code{httpuv::\link[httpuv:stopAllServers]{stopAllServers}} for details.
}
\examples{
beakr1 <- newBeakr()
beakr2 <- newBeakr()
beakr1 \%>\% listen(daemon = TRUE, port = 1234, verbose = TRUE)
beakr2 \%>\% listen(daemon = TRUE, port = 4321, verbose = TRUE)
length(listServers())
stopAllServers()
length(listServers())
}
