% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopServer}
\alias{stopServer}
\title{Stop a beakr instance server}
\usage{
stopServer(beakr = NULL, verbose = FALSE)
}
\arguments{
\item{beakr}{\code{Beakr} instance.}

\item{verbose}{Logical specifying whether to print out details of the
\code{Beakr} instance just stopped.}
}
\value{
None
}
\description{
Stops the server associated with a \code{Beakr} instance,
closing all open connections and unbinding the port.
}
\examples{
library(beakr)

beakr <- newBeakr()

# beakr pipeline
beakr \%>\%

  handleErrors() \%>\%

  listen(daemon = TRUE, verbose = TRUE)

stopServer(beakr, verbose = TRUE)
}
