% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_request}
\alias{test_request}
\title{Test request}
\usage{
test_request(beakr, test_request)
}
\arguments{
\item{beakr}{the beakr instance}

\item{test_request}{the TestRequest instance}
}
\description{
Test request
}
