% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Request.R
\name{TestRequest}
\alias{TestRequest}
\title{TestRequest Class}
\description{
The \code{TestRequest} object is for internal testing only.
}
\seealso{
\code{\link{Response}} and \code{\link{Request}}
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-path}{\code{TestRequest$path()}}
\item \href{#method-method}{\code{TestRequest$method()}}
\item \href{#method-query_string}{\code{TestRequest$query_string()}}
\item \href{#method-body}{\code{TestRequest$body()}}
\item \href{#method-set_header}{\code{TestRequest$set_header()}}
\item \href{#method-get_header}{\code{TestRequest$get_header()}}
\item \href{#method-print}{\code{TestRequest$print()}}
\item \href{#method-clone}{\code{TestRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$path(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-method"></a>}}
\subsection{Method \code{method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$method(method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_string"></a>}}
\subsection{Method \code{query_string()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$query_string(qstring)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-body"></a>}}
\subsection{Method \code{body()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$body(body)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_header"></a>}}
\subsection{Method \code{set_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$set_header(key, value, prefix = "HTTP_")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_header"></a>}}
\subsection{Method \code{get_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$get_header(key, prefix = "HTTP_")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
