% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{websocket}
\alias{websocket}
\title{Websocket Access}
\usage{
websocket(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
Access the instance through websocket. \strong{Experimental.}
}
\examples{
\donttest{
beakr() \%>\%
  websocket('/ws_test', function(req, res, err) 'Websocket Test') \%>\%
  listen()
}
}
