% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{addListener}
\alias{addListener}
\title{Internal function to add listeners}
\usage{
addListener(beakr, FUN, event)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{FUN}{listener function}

\item{event}{the event to listen for, such as "start" or "finish".}
}
\description{
Internal function.
}
