% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beakr.R
\docType{package}
\name{beakr-package}
\alias{beakr-package}
\alias{Beakr-Package}
\title{A minimalist web framework.}
\description{
The \code{beakr} package provides a minimalist web framework for for
developing application programming interfaces in R. The package includes
basic functionality for handling common \code{HTTP} requests.

beakr allows R code to listen for and interact with HTTP and WebSocket clients, so
you can serve web traffic directly out of your R process. beakr relies heavily
on the \href{https://github.com/rstudio/httpuv}{httpuv} package, and therefore
the lower level \href{https://github.com/joyent/libuv}{libuv} and
\href{https://github.com/joyent/http-parser}{http-parser} C libraries.
\code{beakr} is a ground-up rewrite and
continuation of the \code{jug} package developed by Bart Smeets. The
\code{beakr} package is supported and maintained by
\href{http://www.mazamascience.com/}{Mazama Science}.
}
\seealso{
\link{new_beakr}
}
\author{
Hans Martin \email{hans@mazamascience.com}
}
\keyword{package}
