% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beakr}
\alias{beakr}
\title{Create a new Beakr instance}
\usage{
beakr(name = NULL)
}
\arguments{
\item{name}{an optional name assigned to the \emph{Beakr} object.}
}
\description{
Create a \code{Beakr} instance object by calling the top-level
\code{beakr()} function. If \code{name} is not supplied, a random name
will be assigned.
}
