% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beakr}
\alias{beakr}
\title{Start a new beakr instance}
\usage{
beakr()
}
\value{
a beakr class instance
}
\description{
Creates a new beakr instance that can be used to build on with
other functions (middleware).
}
