% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beakr.R
\docType{data}
\name{Beakr}
\alias{Beakr}
\title{Beakr class}
\description{
The \code{Beakr} class defines the server instance utilizing the
\code{httpuv} package. This class defines an interface for the rest of the
\code{beakr} package and is therefore meant to be instantiated.
}
\section{Methods}{


\describe{
  \item{\code{route()}}{
  An instantiated \code{Router} object.
  }
  \item{\code{serverObject()}}{
  The instantiated \code{Server} object.
  }
  \item{\code{appDefinition()}}{
  A method to define the functions or middleware of users application.
  }
  \item{\code{addCollectedMiddleware(collector)}}{
  Adds the users "collected" middleware to the \code{Beakr} defined
  middleware.
  }
  \item{\code{initialize()}}{
  Creates a new \code{Router} object for the \code{route}
  method.
  }
  \item{\code{start(host, port, daemonized)}}{
  Returns a running server. If \code{daemonized = TRUE}, the server will run
  in the background.
  }
  \item{\code{print(...)}}{
  Returns a console output of the instance and its number of middleware
  attached.
  }
}
}

\section{Package details}{


The \code{beakr} package provides a minimal web framework for for developing
lightweight APIs in R. The package includes basic functionality for handling
common \code{HTTP} requests. \code{beakr} is a ground-up rewrite and
continuation of the \code{jug} package developed by Bart Smeets. The
\code{beakr} package is supported and maintained by
\href{http://www.mazamascience.com/}{Mazama Science}.
}

\seealso{
\code{\link{Router}} and \code{\link{Middleware}}
}
\keyword{internal}
