% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cors}
\alias{cors}
\title{Add CORS-headers middleware}
\usage{
cors(
  beakr,
  path = NULL,
  with_methods = c("POST", "GET", "PUT", "OPTIONS", "DELETE", "PATCH"),
  with_origin = "*",
  with_headers = NULL,
  with_credentials = NULL,
  max_age = NULL,
  expose_headers = NULL
)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{a string path.}

\item{with_methods}{set \code{Access-Control-Allow-Methods} response header
that specifies the method or methods allowed when accessing the resource in
response to a preflight request.}

\item{with_origin}{set \code{Access-Control-Allow-Origin} response header
indicating whether the response can be shared with requesting code from the
given origin.}

\item{with_headers}{set \code{Acess-Control-Allow-Headers} to indicate which HTTP
headers can be used during the actual request.}

\item{with_credentials}{set \code{Access-Control-Allow-Credentials} response header.}

\item{max_age}{set \code{Access-Control-Max-Age} in seconds.}

\item{expose_headers}{set \code{Access-Control-Expose-Headers}  indicate which
headers can be exposed as part of the response.}
}
\description{
Add Cross-Origin Resource Sharing (CORS) middleware to the
instance.
}
\details{
CORS is a mechanism that uses additional HTTP headers to tell browsers to
give a web application running at one origin, access to selected resources
from a different origin. A web application executes a cross-origin HTTP
request when it requests a resource that has a different origin
(domain, protocol, or port) from its own.

Additional Info:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS
}
