% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cors}
\alias{cors}
\title{Title}
\usage{
cors(beakr, path = NULL, with_methods = c("POST", "GET", "PUT",
  "OPTIONS", "DELETE", "PATCH"), with_origin = "*",
  with_headers = NULL, with_credentials = NULL, max_age = NULL,
  expose_headers = NULL)
}
\arguments{
\item{expose_headers}{}
}
\description{
Title
}
