% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-middleware.R
\name{cors}
\alias{cors}
\title{Allow Cross-Origin-Requests}
\usage{
cors(
  beakr,
  path = NULL,
  methods = c("GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"),
  origin = "*"
)
}
\description{
Allow CORS-headers.
}
