% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Router.R
\name{.matchPath}
\alias{.matchPath}
\title{Regex path query}
\usage{
.matchPath(pattern, path, ...)
}
\arguments{
\item{pattern}{the string pattern to parse}

\item{path}{the path to match to}

\item{...}{additional parameters}
}
\description{
Regex path query
}
\keyword{internal}
