% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Error.R
\docType{data}
\name{Error}
\alias{Error}
\title{Error class}
\description{
A \code{Error} class object represents the state and handling of instance
or middleware errors.
}
\section{Fields}{


\describe{
  \item{\code{errors}}{
  Returns a list of errors, if any.
  }
  \item{\code{occured}}{
  Returns TRUE if any error has occurred, FALSE otherwise.
  }
}
}

\section{Methods}{


\describe{
  \item{\code{set(err)}}{
  Sets an error.
  }
}
}

\seealso{
\code{\link{Beakr}} and \code{\link{Middleware}}
}
\keyword{internal}
