% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Error.R
\name{Error}
\alias{Error}
\title{Error class}
\description{
A \code{Error} class object represents the state and handling of instance
or middleware errors.
}
\section{Fields}{


\describe{
  \item{\code{errors}}{
  Returns a list of errors, if any.
  }
  \item{\code{occured}}{
  Returns TRUE if any error has occurred, FALSE otherwise.
  }
}
}

\section{Methods}{


\describe{
  \item{\code{set(err)}}{
  Sets an error.
  }
}
}

\seealso{
\code{\link{handler}} and \code{\link{Middleware}}
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set}{\code{Error$set()}}
\item \href{#method-clone}{\code{Error$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$set(err)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
