% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Middleware.R
\name{get}
\alias{get}
\alias{get.Beakr}
\title{GET-binding middleware}
\usage{
get(beakr, ...)

\method{get}{Beakr}(object, ...)
}
\arguments{
\item{...}{}
}
\description{
GET-binding middleware
}
\section{Methods (by class)}{
\itemize{
\item \code{Beakr}: Function to add GET-binding middleware
}}

