% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{get}
\alias{get}
\alias{get.default}
\alias{get.Beakr}
\title{GET-binding middleware}
\usage{
get(beakr, path, ...)

\method{get}{default}(beakr, ...)

\method{get}{Beakr}(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP GET requests to the specified path with the specified callback
functions or middleware.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Base R \code{get} function.

\item \code{Beakr}: Beakr middleware function.
}}

\examples{
\dontrun{
# Create a beakr instance
server <- beakr() \%>\%
  get("/", function(req, res, err) { "Hello, world!" })
# Listen for HTTP/WebSocket requests
listen(server, daemonized = TRUE)
#' # In terminal:
$ curl http://127.0.0.1:8080/
> Hello, world!
}
}
