% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-middleware.R
\name{handleErrors}
\alias{handleErrors}
\title{Error-handling middleware}
\usage{
handleErrors(beakr = NULL, FUN = jsonError)
}
\arguments{
\item{beakr}{Beakr instance}

\item{FUN}{a function to handle the error response}
}
\value{
A \code{Beakr} instance with added middleware.
}
\description{
This default error handler should be added
at the end of the beakr pipeline, right before \code{listen()}. Errors
generated by any previous step will be returned within a JSON wrapper.

The general structure for a stand-alone executable script with a
\pkg{beakr} web service typically looks like this:

\preformatted{
newBeakr() \%>\%

 # Respond to GET requests at the "/hi" route
 httpGET(path = "/hi", function(req, res, err) {
   print("Hello, World!")
 }) /%>/%

 # Respond to GET requests at the "/bye" route
 httpGET(path = "/bye", function(req, res, err) {
   print("Farewell, my friends.")
 }) \%>\%

 handleErrors() \%>\%

 listen()
}
}
