% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handleErrors}
\alias{handleErrors}
\title{Instance Error Handling}
\usage{
handleErrors(beakr, path)
}
\arguments{
\item{beakr}{a beakr instance}

\item{path}{string representing a relative path for which the middleware
is invoked.}
}
\description{
An error handling middleware for beakr instances.
}
\details{
\code{beakr} comes with a default error handler. This default error-handling
middleware function is added at the end of the middleware function stack.
The Errors are handled via JSON output.
}
