% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{httpPOST}
\alias{httpPOST}
\title{POST-binding middleware}
\usage{
httpPOST(beakr, path = NULL, ...)
}
\arguments{
\item{beakr}{\code{Beakr} instance or \code{NULL}.}

\item{path}{String representing a relative path for which the middleware
is invoked.}

\item{...}{Additional middleware/functions.}
}
\value{
A \code{Beakr} object with added middleware.
}
\description{
Routes HTTP POST requests to the specified path with the
specified callback functions or middleware.
}
\examples{
\dontrun{
# Create an new Beakr instance
beakr <- newBeakr()

# Create a simple beakr pipeline
beakr \%>\%
  httpPOST("/", function(req, res, err) {
    return("Successful POST request!\n")
  }) \%>\%
  listen(host = '127.0.0.1', port = 12518, daemon = TRUE)

# ------------------------------------------------------------
# IN A TERMINAL:
#  curl -X POST http://127.0.0.1:12518/
# > Successful POST request!
# ------------------------------------------------------------

# Kill the beakr instance
kill(beakr)
}
}
