% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kill}
\alias{kill}
\title{Stop the beakr instance}
\usage{
kill(beakr)
}
\arguments{
\item{beakr}{a beakr instance.}
}
\description{
Stops an active instance when given a beakr object. This closes
all open connections for the instance and unbinds the port.
}
\examples{
\donttest{
beakr <- newBeakr()
startBeakr(beakr, daemonized = TRUE)
kill(beakr)
}

}
