% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{killall}
\alias{killall}
\title{Stop all instances}
\usage{
killall()
}
\description{
Stops all instances that have been activated by \code{\link{startBeakr}} in
the session.
}
\seealso{
\code{\link{kill}} and \code{\link{startBeakr}}
}
