% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listen}
\alias{listen}
\title{Listen for connections on a beakr instance}
\usage{
listen(beakr, host, port, daemon)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{host}{a string that is a valid IPv4 or IPv6 address to listen on.
Defaults to the local host ("127.0.0.1").}

\item{port}{a number or integer that indicates the port to listen on. Default
is a port opened on 8080.}

\item{daemon}{run the instance in the background, the default is FALSE.}
}
\value{
A `beakr` App object.
}
\description{
Binds and listens for connections on the specified host and port.
}
\details{
\code{listen()} binds the port and listens for connections on a thread.
The thread handles the I/O, and when it receives a HTTP request, it
will schedule a call to the user-defined middleware and handle the
request.

If the daemon boolean value is set to \code{TRUE}, \code{listen()} binds
the specified port and listens for connections on a thread running in the
background.

See the \code{httpuv} package for more information.
}
\examples{
\dontrun{
# Run in foreground
newBeakr() \%>\%
  httpGET("/", function(req, res, err) {
    return("Successful GET request!\n")
  }) \%>\%
  listen()

# Run in background
#' newBeakr() \%>\%
  httpGET("/", function(req, res, err) {
    return("Successful GET request!\n")
  }) \%>\%
  listen(daemon = TRUE)
}
}
