% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listener.R
\docType{data}
\name{Listener}
\alias{Listener}
\title{Listener class}
\description{
The \code{Listener} class listens...to things.
}
\section{Fields}{


\describe{
  \item{\code{FUN}}{
  Returns function response.
  }
  \item{\code{event}}{
  Returns event type.
  }
}
}

\section{Methods}{


\describe{
  \item{\code{initialize(FUN, event)}}{
  Sets instance object function and event state.
  }
}
}

\seealso{
\code{\link{Router}} and \code{\link{Error}}
}
\keyword{internal}
