% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newBeakr}
\alias{newBeakr}
\title{Create a new Beakr instance}
\usage{
newBeakr(name = NULL)
}
\arguments{
\item{name}{an optional name assigned to the \emph{Beakr} object.}
}
\value{
A new and empty `beakr` App.
}
\description{
Create a \emph{Beakr} instance object by calling the top-level
\code{newBeakr()} function. If \code{name} is not supplied, a random name
will be assigned.
}
\examples{
\dontrun{
# Standard
beakr <- newBeakr()
listen(beakr)
# Pipe
newBeakr() \%>\% listen()
}
}
