% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{parseParameters}
\alias{parseParameters}
\title{Parse the parameters passed by the req}
\usage{
parseParameters(req, body, query, type)
}
\arguments{
\item{req}{an HTTP req}

\item{body}{a body text string}

\item{query}{a url-encoded query string}

\item{type}{a media mime type (a.k.a. Multipurpose Internet Mail Extensions
or MIME type).}
}
\description{
Parse the parameters passed by the req
}
