% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseParameters}
\alias{parseParameters}
\title{Parse the parameters passed by the request}
\usage{
parseParameters(request, body, query, type)
}
\arguments{
\item{request}{an HTTP request}

\item{body}{a body text string}

\item{query}{a url-encoded query string}

\item{type}{a media type (a.k.a. Multipurpose Internet Mail Extensions or MIME type)}
}
\value{
parsed parameters list
}
\description{
Parse the parameters passed by the request
}
