% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{put}
\alias{put}
\title{PUT-binding middleware}
\usage{
put(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Routes HTTP PUT requests to the specified path with the specified callback
functions or middleware.
}
