% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Request.R
\docType{data}
\name{Request}
\alias{Request}
\title{Request Object Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Request
}
\description{
Request Object Class
}
\keyword{datasets}
