% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\docType{data}
\name{Response}
\alias{Response}
\title{Response Object Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Response
}
\description{
Response Object Class
}
\keyword{datasets}
