% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{routeMiddleware}
\alias{routeMiddleware}
\title{Internal function to add middleware}
\usage{
routeMiddleware(beakr, FUN, path = NULL, method = NULL,
  websocket = FALSE)
}
\arguments{
\item{beakr}{a beakr instance.}

\item{FUN}{the function to route middleware}

\item{path}{the path to route the middleware}

\item{method}{the HTTP method to employ}

\item{websocket}{boolean, TRUE if websocket.}
}
\description{
Internal function to add middleware
}
\keyword{internal}
