% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Router.R
\docType{data}
\name{Router}
\alias{Router}
\title{Router Class}
\description{
The \code{Router} object represents the handling of routing and middleware
(such as get, put, post, and so on). Once a \code{Router} object is
instantiated, middleware and HTTP method routes can be added. The top level
\code{Beakr} object initializes with the creation of a \code{Router} object.
}
\section{Fields}{


\describe{
  \item{\code{middleware}}{
  A list of specified middleware function or functions.
  }
  \item{\code{listeners}}{
  A list of specified listeners.
  }
}
}

\section{Methods}{


\describe{
\item{\code{addMiddleware(middlware)}}{
  A method to add middleware function(s) to \code{middleware}.
  }
  \item{\code{addListener(listener)}}{
  A method to add listeners to \code{listeners}.
  }
  \item{\code{processEvent(event, ...)}}{
  Processes the event heard by the \code{Listener}.
  }
  \item{\code{invoke(req, websocket_msg, websocket_binary)}}{
  This method is used to create the a request-response cycle objects of the
  provided middleware.
  }
}
}

\keyword{internal}
