% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serveStaticFiles}
\alias{serveStaticFiles}
\title{Serve static files}
\usage{
serveStaticFiles(beakr, path = NULL, file = NULL)
}
\arguments{
\item{beakr}{\code{Beakr} instance or \code{NULL}.}

\item{path}{String representing a path to a directory from which static
files are served.}

\item{file}{a string representing the path to the file that is to be served.}
}
\value{
A `beakr` App object with added middleware.
}
\description{
Binds to GET requests that aren't handled by specified paths.
Should support all filetypes; returns image and octet-stream types as a raw
string.
}
\details{
Serve static files from the host machine. Currently supports images,
PDF, JSON, HTML, and raw text. The static file will be served on the \code{path}
specified in addition to the file name and extension. For example, an image
\code{example_dir/path_to_pictures/pic.png} will be served on the URL extension
\code{/path/pic.png}.
}
\examples{
\dontrun{
path_to_file <- "example_dir/path_to_pictures/pic.png"
newBeakr() \%>\%
  static(path = 'readme/', file = path_to_file) \%>\%
  listen()
}
}
