% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestRequest.R
\docType{data}
\name{TestRequest}
\alias{TestRequest}
\title{Generate request for testing purposes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TestRequest
}
\description{
solely used for testing purposes
}
\keyword{datasets}
