% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{webSocket}
\alias{webSocket}
\title{Websocket Access}
\usage{
webSocket(beakr, path, ...)
}
\arguments{
\item{beakr}{a beakr instance}

\item{path}{string representing a relative path for which the middleware
is invoked.}

\item{...}{additional middleware/functions.}
}
\description{
Access the instance through websocket. \strong{Experimental.}
}
